+++ test/hocsMultipleProps.spec.js
@@ -6,7 +6,12 @@ import { render, screen } from "@testing-library/react";
 import { providers } from "@data-provider/core";
 import sinon from "sinon";
 
-import { withDataLoadedError, withDataLoadingError, withDataLoadingErrorComponents } from "../src";
+import {
+  withDataLoadedError,
+  withDataLoadingError,
+  withDataLoadingErrorComponents,
+  withDataLoadedErrorComponents,
+} from "../src";
 
 import MockProvider from "./MockProvider";
 import { BOOKS, BOOKS_ID, LOADING_ID, ERROR_ID } from "./constants";
@@ -36,9 +41,8 @@ describe("HOCs providing multiple props", () => {
 
   describe("withDataLoadedError", () => {
     beforeEach(() => {
-      provider = new MockProvider(`${BOOKS_ID}-1`, {
-        data: BOOKS,
-      });
+      provider.resetState();
+      provider.cleanCache();
       BooksComponent = ({ data, loaded, error }) => {
         return <Books error={error} books={data} loading={!loaded} />;
       };
@@ -108,9 +112,8 @@ describe("HOCs providing multiple props", () => {
 
   describe("withDataLoadedError using custom props", () => {
     beforeEach(() => {
-      provider = new MockProvider(`${BOOKS_ID}-2`, {
-        data: BOOKS,
-      });
+      provider.resetState();
+      provider.cleanCache();
       BooksComponent = ({ books, booksAreLoaded, booksError }) => {
         return <Books error={booksError} books={books} loading={!booksAreLoaded} />;
       };
@@ -160,9 +163,8 @@ describe("HOCs providing multiple props", () => {
 
   describe("withDataLoadedError using data custom prop", () => {
     beforeEach(() => {
-      provider = new MockProvider(`${BOOKS_ID}-3`, {
-        data: BOOKS,
-      });
+      provider.resetState();
+      provider.cleanCache();
       BooksComponent = ({ books, loaded, error }) => {
         return <Books error={error} books={books} loading={!loaded} />;
       };
@@ -587,6 +589,247 @@ describe("HOCs providing multiple props", () => {
 
     it("should return null when provider throws an error and no error component is provided", async () => {
       BooksConnectedComponent = withDataLoadingErrorComponents(provider)(
+        BooksComponent,
+        CustomLoadingComponent
+      );
+
+      Component = () => (
+        <ReduxProvider>
+          <BooksConnectedComponent />
+        </ReduxProvider>
+      );
+      provider.error = new Error();
+      render(<Component />);
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+      expect(screen.queryByTestId(CUSTOM_ERROR_ID)).not.toBeInTheDocument();
+      expect(screen.queryByTestId(BOOKS_ID)).not.toBeInTheDocument();
+    });
+  });
+
+  describe("withDataLoadedErrorComponents", () => {
+    const CUSTOM_LOADING_ID = `${LOADING_ID}-custom`;
+    const CUSTOM_ERROR_ID = `${ERROR_ID}-custom`;
+    let CustomLoadingComponent, CustomErrorComponent;
+
+    beforeEach(() => {
+      provider.resetState();
+      provider.cleanCache();
+      CustomLoadingComponent = ({ loaded }) => {
+        if (loaded) {
+          return null;
+        }
+        return <div data-testid={CUSTOM_LOADING_ID}>Loading</div>;
+      };
+
+      CustomErrorComponent = ({ error = {} }) => {
+        return <div data-testid={CUSTOM_ERROR_ID}>{error.message}</div>;
+      };
+
+      BooksComponent = ({ data }) => {
+        return <Books books={data} />;
+      };
+
+      BooksConnectedComponent = withDataLoadedErrorComponents(provider)(
+        BooksComponent,
+        CustomLoadingComponent,
+        CustomErrorComponent
+      );
+
+      Component = () => (
+        <ReduxProvider>
+          <BooksConnectedComponent />
+        </ReduxProvider>
+      );
+    });
+
+    it("should wrap displayName of the component", async () => {
+      expect(BooksConnectedComponent.displayName).toEqual(
+        "WithDataLoadedErrorComponentsBooksComponent"
+      );
+    });
+
+    it("loading should be true when provider is loading first time and false when finish", async () => {
+      render(<Component />);
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).toBeInTheDocument();
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+    });
+
+    it("loading should not change when provider cache is cleaned", async () => {
+      render(<Component />);
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+      provider.cleanCache();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+    });
+
+    it("should return null when provider is loading and no loading component is provided", async () => {
+      BooksConnectedComponent = withDataLoadedErrorComponents(provider)(BooksComponent);
+
+      Component = () => (
+        <ReduxProvider>
+          <BooksConnectedComponent />
+        </ReduxProvider>
+      );
+      render(<Component />);
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+      expect(screen.queryByTestId(CUSTOM_ERROR_ID)).not.toBeInTheDocument();
+      expect(screen.queryByTestId(BOOKS_ID)).not.toBeInTheDocument();
+    });
+
+    it("should pass data to the component", async () => {
+      const bookTitle = "Animal Farm";
+      render(<Component />);
+      await wait();
+      expect(screen.getByText(bookTitle)).toBeInTheDocument();
+    });
+
+    it("should refresh data", async () => {
+      const TEST_ID = "book-2";
+      render(<Component />);
+      await wait();
+      expect(screen.queryByTestId(TEST_ID)).toBeInTheDocument();
+      provider.delete(2);
+      await wait();
+      expect(screen.queryByTestId(TEST_ID)).not.toBeInTheDocument();
+    });
+
+    it("error should be null when provider does not throw error", async () => {
+      render(<Component />);
+      expect(screen.queryByTestId(CUSTOM_ERROR_ID)).not.toBeInTheDocument();
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_ERROR_ID)).not.toBeInTheDocument();
+    });
+
+    it("should return error when provider throws error", async () => {
+      const ERROR_MESSAGE = "Foo error";
+      provider.error = new Error(ERROR_MESSAGE);
+      render(<Component />);
+      await wait();
+      expect(screen.getByText(ERROR_MESSAGE)).toBeInTheDocument();
+    });
+
+    it("should return null when provider throws an error and no error component is provided", async () => {
+      BooksConnectedComponent = withDataLoadedErrorComponents(provider)(
+        BooksComponent,
+        CustomLoadingComponent
+      );
+
+      Component = () => (
+        <ReduxProvider>
+          <BooksConnectedComponent />
+        </ReduxProvider>
+      );
+      provider.error = new Error();
+      render(<Component />);
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+      expect(screen.queryByTestId(CUSTOM_ERROR_ID)).not.toBeInTheDocument();
+      expect(screen.queryByTestId(BOOKS_ID)).not.toBeInTheDocument();
+    });
+  });
+
+  describe("withDataLoadedErrorComponents using custom prop", () => {
+    const CUSTOM_LOADING_ID = `${LOADING_ID}-custom`;
+    const CUSTOM_ERROR_ID = `${ERROR_ID}-custom`;
+    let CustomLoadingComponent, CustomErrorComponent;
+
+    beforeEach(() => {
+      provider = new MockProvider(BOOKS_ID, {
+        data: BOOKS,
+      });
+      CustomLoadingComponent = ({ booksAreLoaded }) => {
+        if (booksAreLoaded) {
+          return null;
+        }
+        return <div data-testid={CUSTOM_LOADING_ID}>Loading</div>;
+      };
+
+      CustomErrorComponent = ({ booksError = {} }) => {
+        return <div data-testid={CUSTOM_ERROR_ID}>{booksError.message}</div>;
+      };
+
+      BooksConnectedComponent = withDataLoadedErrorComponents(provider, [
+        "books",
+        "booksAreLoaded",
+        "booksError",
+      ])(Books, CustomLoadingComponent, CustomErrorComponent);
+
+      Component = () => (
+        <ReduxProvider>
+          <BooksConnectedComponent />
+        </ReduxProvider>
+      );
+    });
+
+    it("loading should be true when provider is loading and false when finish", async () => {
+      render(<Component />);
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).toBeInTheDocument();
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+    });
+
+    it("loading should not change when provider cache is cleaned", async () => {
+      render(<Component />);
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+      provider.cleanCache();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+    });
+
+    it("should return null when provider is loading and no loading component is provided", async () => {
+      BooksConnectedComponent = withDataLoadedErrorComponents(provider)(BooksComponent);
+
+      Component = () => (
+        <ReduxProvider>
+          <BooksConnectedComponent />
+        </ReduxProvider>
+      );
+      render(<Component />);
+      expect(screen.queryByTestId(CUSTOM_LOADING_ID)).not.toBeInTheDocument();
+      expect(screen.queryByTestId(CUSTOM_ERROR_ID)).not.toBeInTheDocument();
+      expect(screen.queryByTestId(BOOKS_ID)).not.toBeInTheDocument();
+    });
+
+    it("should pass data to the component", async () => {
+      const bookTitle = "Animal Farm";
+      render(<Component />);
+      await wait();
+      expect(screen.getByText(bookTitle)).toBeInTheDocument();
+    });
+
+    it("should refresh data", async () => {
+      const TEST_ID = "book-2";
+      render(<Component />);
+      await wait();
+      expect(screen.queryByTestId(TEST_ID)).toBeInTheDocument();
+      provider.delete(2);
+      await wait();
+      expect(screen.queryByTestId(TEST_ID)).not.toBeInTheDocument();
+    });
+
+    it("error should be null when provider does not throw error", async () => {
+      render(<Component />);
+      expect(screen.queryByTestId(CUSTOM_ERROR_ID)).not.toBeInTheDocument();
+      await wait();
+      expect(screen.queryByTestId(CUSTOM_ERROR_ID)).not.toBeInTheDocument();
+    });
+
+    it("should return error when provider throws error", async () => {
+      const ERROR_MESSAGE = "Foo error";
+      provider.error = new Error(ERROR_MESSAGE);
+      render(<Component />);
+      await wait();
+      expect(screen.getByText(ERROR_MESSAGE)).toBeInTheDocument();
+    });
+
+    it("should return null when provider throws an error and no error component is provided", async () => {
+      BooksConnectedComponent = withDataLoadedErrorComponents(provider)(
         BooksComponent,
         CustomLoadingComponent
       );